Hw5

+++ feature/todo/src/main/java/com/cleverpumpkin/todo/presentation/screens/todo_detail_screen/TodoDetailViewModel.kt
@@ -2,9 +2,9 @@ package com.cleverpumpkin.todo.presentation.screens.todo_detail_screen
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.cleverpumpkin.cor.id_handlers.IdGenerator
-import com.cleverpumpkin.cor.presentation.navigation.NavArgs
-import com.cleverpumpkin.networ.domain.response_wrapper.Response
+import com.cleverpumpkin.core.id_handlers.IdGenerator
+import com.cleverpumpkin.core.presentation.navigation.NavArgs
+import com.cleverpumpkin.network.domain.response_wrapper.Response
 import com.cleverpumpkin.todo.domain.todo_model.Importance
 import com.cleverpumpkin.todo.domain.todo_model.TodoItem
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -94,7 +94,9 @@ class TodoDetailViewModel @Inject constructor(
     }
 
     fun deleteItem() = viewModelScope.launch {
-        processResponse(repository.deleteTodoItem(createTodo()))
+        if (_uiState.value.id.isNotEmpty()) {
+            processResponse(repository.deleteTodoItem(createTodo()))
+        }
     }
 
     private fun processResponse(response: Response<*>, onSuccessAction: (() -> Unit)? = null) {
